Fixed issue with search function
Initially if you were trying to search for something else while already viewing 2 out of 100 pages on a search it would either bug out or not work or save that same page for the next search. It has now been fixed so that when you make a new search the active page resets to 1.